+++ backend/src/main/java/com/dev/backend/model/Reservation.java
@@ -9,7 +9,7 @@ public class Reservation {
 
     @Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
+	private int id;
 
 	@Column(name = "status", nullable = false)
 	private int status;
@@ -38,10 +38,10 @@ public class Reservation {
         this.client = client;
 	}
 	
-	public long getId() {
+	public int getId() {
 		return id;
 	}
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	
@@ -70,7 +70,7 @@ public class Reservation {
 		return book;
 	}
 
-	public Long getBookId() {
+	public int getBookId() {
 		return book.getId();
 	}
 
@@ -82,7 +82,7 @@ public class Reservation {
 		return client;
 	}
 
-	public Long getClientId() {
+	public int getClientId() {
 		return client.getId();
 	}
